test: share lotus-performance stateful fakes across suites
